functions/auto_approve/requirements.txt
google-cloud-datastore==1.8.0
google-python-cloud-debugger
jwkaas==0.3
python-dateutil==2.6.0
